Use a common interface for OptionProps
Previously there were two types named `OptionProps`, this merges them
into one.
